Use Bower to track dependencies
[.bowerrc](.bowerrc) configures the directory for bower to install dependencies, in this case `app/scripts/vendor/`. Bower will create that directory upon calling `$ bower install` from the project root.

Since we are now using bower to install dependencies _including_ requirejs, we must update the path to require.js, to its new home in the vendor folder. This path is adjusted in the `<script>` tag in [index.html](app/index.html).

The path to `jquery` is also updated in [app.js](app/scripts/app.js).

Finally, the `paths` config in [bower.json](bower.json) is updated so that `jquery` points to new install location from bower.

Inspiration for this setup was taken from a [tuts+ template](http://code.tutsplus.com/tutorials/a-requirejs-backbone-and-bower-starter-template--net-29211) for setting up requirejs, backbone, and bower.