Add Backbone
There are amd-compatible versions of underscore and backbone available, which gulp will find by using the `-amd` suffix, as in [bower.json](bower.json):

```json
{
  "dependencies": {
    "underscore-amd": null,
    "backbone-amd": null
  }
}
```

That is the main notable thing about adding backbone. It is given a name in [app/scripts/main.js] just like jquery so that it can easily be `require`d.

The structure of these changes is inspired by the same [tuts+ tutorial](http://code.tutsplus.com/tutorials/a-requirejs-backbone-and-bower-starter-template--net-29211) as the last.